+++ src/app/beta/[...page]/main_page.tsx
